refactor: move middleware to pkg/observability
